+++ tests/functional_tests/expected_values/bitly.py
@@ -36,6 +36,7 @@ CONFIG = {
         'gender': None,
         'link': LINK,
         'locale': None,
+        'location': None,
         'phone': None,
         'picture': PICTURE,
         'postal_code': None,