+++ tests/functional_tests/expected_values/openid_yahoo.py
@@ -1,7 +1,4 @@
-import re
-
 import fixtures
-import constants
 from authomatic.providers import openid
 
 conf = fixtures.get_configuration('openid_yahoo')
@@ -27,6 +24,7 @@ CONFIG = {
         'gender': None,
         'link': None,
         'locale': None,
+        'location': None,
         'phone': None,
         'picture': None,
         'postal_code': None,