+++ tests/functional_tests/expected_values/reddit.py
@@ -27,6 +27,7 @@ CONFIG = {
         'gender': None,
         'link': None,
         'locale': None,
+        'location': None,
         'phone': None,
         'picture': None,
         'postal_code': None,